backend/flaskapp/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
PyMySQL==1.1.0
cryptography
mysql-connector-python==9.1.0